Use Debian for builder images
The version of node-gyp that this project uses requires Python 3.9, which is not
available on the version of Alpine linux that we were using. This change moves
the images that require node-gyp to Debian Bullseye. I have verified that these
new images can be built but I've been unable to test them more thouroughly due
to other setup issues.